Fixing issues
